Refactor 'osu.data' package as 'osu.conversion'
