+++ src/com/reco1l/osu/Multiplayer.kt
@@ -5,8 +5,8 @@ import android.util.Log
 import com.reco1l.api.ibancho.RoomAPI
 import com.reco1l.api.ibancho.data.Room
 import com.reco1l.api.ibancho.data.RoomPlayer
-import com.reco1l.osu.data.jsonToScoreboardItem
-import com.reco1l.osu.data.jsonToStatistic
+import com.reco1l.osu.conversion.jsonToScoreboardItem
+import com.reco1l.osu.conversion.jsonToStatistic
 import com.reco1l.osu.ui.multiplayer.RoomScene
 import com.reco1l.toolkt.kotlin.formatTimeMilliseconds
 import com.reco1l.toolkt.kotlin.fromDate
